# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentação da API Usuários",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar todos os usuários",
                "description": "Retorna a lista de usuários.",
                "operationId": "600ec787a82d08ec298424fd97dd33d9",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Cadastrar um novo usuário",
                "description": "Realiza o cadastro de um novo usuário.",
                "operationId": "12af6b3d715bf3099de301b50a13d0da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "nome",
                                    "data_nascimento"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "unsignedBigInteger",
                                        "example": 12345678900
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "Name"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário cadastrado com sucesso"
                    },
                    "409": {
                        "description": "Não foi possível cadastrar o usuário"
                    }
                }
            }
        },
        "/api/usuarios/{cpf}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Buscar um usuário pelo CPF",
                "description": "Retorna as informações de um usuário baseado no CPF fornecido.",
                "operationId": "83bf07375fb378994bdf4865c788a481",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "CPF do usuário a ser buscado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 12345678900
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Usuários",
            "description": "Usuários"
        }
    ]
}